requirements.txt
# requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
PyJWT==2.1.0
bcrypt==3.2.0
python-dotenv==0.19.0